SW-4631 Refactor planting site view to accept a generic type
- generic type is lowest common denominator between sites and draft sites
- most of the planting site specific data like observations is not affected
- next step is to leverage this view to show draft sites which will have similar information
