+++ src/services/MapService.ts
@@ -1,5 +1,5 @@
 import { paths } from 'src/api/types/generated-schema';
-import { MultiPolygon, PlantingSite } from 'src/types/Tracking';
+import { MultiPolygon, PlantingSite, SiteDetails } from 'src/types/Tracking';
 import { MapBoundingBox, MapData, MapEntity, MapGeometry, MapSourceBaseData } from 'src/types/Map';
 import HttpService, { Response } from './HttpService';
 import {
@@ -141,7 +141,7 @@ const getPolygons = (boundary?: MultiPolygon): MapGeometry => {
 /**
  * Transform planting site geometry data into UI model
  */
-const extractPlantingSite = (site: PlantingSite): MapSourceBaseData => {
+const extractPlantingSite = (site: SiteDetails): MapSourceBaseData => {
   const { id, name, description, boundary } = site;
 
   return {
@@ -159,7 +159,7 @@ const extractPlantingSite = (site: PlantingSite): MapSourceBaseData => {
 /**
  * Transform zones geometry data into UI model
  */
-const extractPlantingZones = (site: PlantingSite): MapSourceBaseData => {
+const extractPlantingZones = (site: SiteDetails): MapSourceBaseData => {
   const zonesData =
     site.plantingZones?.map((zone) => {
       const { id, name, boundary } = zone;
@@ -179,7 +179,7 @@ const extractPlantingZones = (site: PlantingSite): MapSourceBaseData => {
 /**
  * Transform subzones geometry data into UI model
  */
-const extractSubzones = (site: PlantingSite): MapSourceBaseData => {
+const extractSubzones = (site: SiteDetails): MapSourceBaseData => {
   const allPlantingSubzonesData =
     site.plantingZones?.flatMap((zone) => {
       const { plantingSubzones } = zone;
